// src/cartService/proto/cart/Product.ts
// Original file: cart.proto


export interface Product {
  'productName'?: (string);
  'productPrice'?: (number);
  'productUUID'?: (string);
  'productDescription'?: (string);
  'productImage'?: (string);
  'productQuantity'?: (number);
}

export interface Product__Output {
  'productName': (string);
  'productPrice': (number);
  'productUUID': (string);
  'productDescription': (string);
  'productImage': (string);
  'productQuantity': (number);
}